move the files to view object folder
