+++ backend/StudentService/src/main/java/com/example/studentservice/Vo/Status.java
@@ -1,4 +1,4 @@
-package com.example.studentservice.Model;
+package com.example.studentservice.Vo;
 
 public enum Status {
     NEW, APPLIED, APPROVED, REJECTED, IN_PROGRESS, COMPLETED, CANCELLED
