+++ backend/StudentService/src/main/java/com/example/studentservice/Vo/UserCredential.java
@@ -1,9 +1,5 @@
-package com.example.studentservice.Model;
+package com.example.studentservice.Vo;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
