+++ backend/StudentService/src/main/java/com/example/studentservice/Vo/UserRole.java
@@ -1,4 +1,4 @@
-package com.example.studentservice.Model;
+package com.example.studentservice.Vo;
 
 public enum UserRole {
     FACULTY,
